python/requirements.txt
numpy
Pillow
pyinstaller